Pivotal ID # 186657310: Task Completion Publish Message
Remove unnecessary code
