+++ submission/submission-core/src/main/kotlin/ac/uk/ebi/biostd/submission/config/SubmissionConfig.kt
@@ -7,12 +7,10 @@ import ac.uk.ebi.biostd.persistence.common.service.SubmissionPersistenceService
 import ac.uk.ebi.biostd.persistence.filesystem.api.FileStorageService
 import ac.uk.ebi.biostd.submission.domain.extended.ExtSubmissionQueryService
 import ac.uk.ebi.biostd.submission.domain.extended.ExtSubmissionService
-import ac.uk.ebi.biostd.submission.domain.request.SubmissionStagesHandler
 import ac.uk.ebi.biostd.submission.domain.submission.SubmissionQueryService
 import ac.uk.ebi.biostd.submission.domain.submission.SubmissionService
 import ac.uk.ebi.biostd.submission.domain.submission.SubmissionSubmitter
 import ac.uk.ebi.biostd.submission.domain.submitter.ExtSubmissionSubmitter
-import ac.uk.ebi.biostd.submission.stats.SubmissionStatsService
 import ebi.ac.uk.extended.mapping.to.ToFileListMapper
 import ebi.ac.uk.extended.mapping.to.ToSubmissionMapper
 import ebi.ac.uk.security.integration.components.ISecurityQueryService
@@ -59,13 +57,6 @@ class SubmissionConfig {
         submissionPersistenceService,
         fileStorageService
     )
-
-    @Bean
-    fun submissionStagesHandler(
-        statsService: SubmissionStatsService,
-        extSubmissionSubmitter: ExtSubmissionSubmitter,
-        eventsPublisherService: EventsPublisherService,
-    ): SubmissionStagesHandler = SubmissionStagesHandler(statsService, extSubmissionSubmitter, eventsPublisherService)
 
     @Bean
     fun extSubmissionQueryService(
