+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/submission/domain/service/SubmissionMessageListener.kt
@@ -1,7 +1,8 @@
 package ac.uk.ebi.biostd.submission.domain.service
 
 import ac.uk.ebi.biostd.common.config.LISTENER_FACTORY_NAME
-import ac.uk.ebi.biostd.submission.domain.request.SubmissionStagesHandler
+import ac.uk.ebi.biostd.submission.domain.submitter.ExtSubmissionSubmitter
+import ac.uk.ebi.biostd.submission.stats.SubmissionStatsService
 import ebi.ac.uk.extended.events.RequestCheckedReleased
 import ebi.ac.uk.extended.events.RequestCleaned
 import ebi.ac.uk.extended.events.RequestCreated
@@ -9,70 +10,100 @@ import ebi.ac.uk.extended.events.RequestFilesCopied
 import ebi.ac.uk.extended.events.RequestFinalized
 import ebi.ac.uk.extended.events.RequestIndexed
 import ebi.ac.uk.extended.events.RequestLoaded
+import ebi.ac.uk.extended.events.RequestMessage
 import ebi.ac.uk.extended.events.RequestPersisted
+import kotlinx.coroutines.runBlocking
 import mu.KotlinLogging
 import org.springframework.amqp.rabbit.annotation.RabbitHandler
 import org.springframework.amqp.rabbit.annotation.RabbitListener
+import uk.ac.ebi.events.service.EventsPublisherService
 
 private val logger = KotlinLogging.logger {}
 
 @RabbitListener(queues = ["\${app.notifications.requestQueue}"], containerFactory = LISTENER_FACTORY_NAME)
 class SubmissionMessageListener(
-    private val stagesHandler: SubmissionStagesHandler,
+    private val statsService: SubmissionStatsService,
+    private val submissionSubmitter: ExtSubmissionSubmitter,
+    private val eventsPublisherService: EventsPublisherService,
 ) {
     @RabbitHandler
     fun indexRequest(rqt: RequestCreated) {
-        val (accNo, version) = rqt
-        logger.info { "$accNo, Received index message for submission $accNo, version: $version" }
-        stagesHandler.indexRequest(rqt)
+        processSafely(rqt) {
+            val (accNo, version) = rqt
+            logger.info { "$accNo, Received index message for submission $accNo, version: $version" }
+            submissionSubmitter.indexRequest(accNo, version)
+        }
     }
 
     @RabbitHandler
     fun loadRequest(rqt: RequestIndexed) {
-        val (accNo, version) = rqt
-        logger.info { "$accNo, Received load message for submission $accNo, version: $version" }
-        stagesHandler.loadRequest(rqt)
+        processSafely(rqt) {
+            val (accNo, version) = rqt
+            logger.info { "$accNo, Received load message for submission $accNo, version: $version" }
+            submissionSubmitter.loadRequest(accNo, version)
+        }
     }
 
     @RabbitHandler
     fun cleanRequest(rqt: RequestLoaded) {
-        val (accNo, version) = rqt
-        logger.info { "$accNo, Received clean message for submission $accNo, version: $version" }
-        stagesHandler.cleanRequest(rqt)
+        processSafely(rqt) {
+            val (accNo, version) = rqt
+            logger.info { "$accNo, Received clean message for submission $accNo, version: $version" }
+            submissionSubmitter.cleanRequest(accNo, version)
+        }
     }
 
     @RabbitHandler
     fun copyRequestFiles(rqt: RequestCleaned) {
-        val (accNo, version) = rqt
-        logger.info { "$accNo, Received persist files message for submission $accNo, version: $version" }
-        stagesHandler.copyRequestFiles(rqt)
+        processSafely(rqt) {
+            val (accNo, version) = rqt
+            logger.info { "$accNo, Received persist files message for submission $accNo, version: $version" }
+            submissionSubmitter.processRequest(accNo, version)
+        }
     }
 
     @RabbitHandler
     fun checkReleased(rqt: RequestFilesCopied) {
-        val (accNo, version) = rqt
-        logger.info { "$accNo, Received check release status message for submission $accNo, version: $version" }
-        stagesHandler.checkReleased(rqt)
+        processSafely(rqt) {
+            val (accNo, version) = rqt
+            logger.info { "$accNo, Received check release status message for submission $accNo, version: $version" }
+            submissionSubmitter.checkReleased(accNo, version)
+        }
     }
 
     @RabbitHandler
     fun saveSubmission(rqt: RequestCheckedReleased) {
-        val (accNo, version) = rqt
-        logger.info { "$accNo, Received save submission message for submission $accNo, version: $version" }
-        stagesHandler.saveSubmission(rqt)
+        processSafely(rqt) {
+            val (accNo, version) = rqt
+            logger.info { "$accNo, Received save submission message for submission $accNo, version: $version" }
+            submissionSubmitter.saveRequest(accNo, version)
+        }
     }
 
     @RabbitHandler
     fun finalizeRequest(rqt: RequestPersisted) {
-        val (accNo, version) = rqt
-        logger.info { "$accNo, Received finalize submission message for submission $accNo, version: $version" }
-        stagesHandler.finalizeRequest(rqt)
+        processSafely(rqt) {
+            val (accNo, version) = rqt
+            logger.info { "$accNo, Received finalize submission message for submission $accNo, version: $version" }
+            submissionSubmitter.finalizeRequest(accNo, version)
+        }
     }
 
     @RabbitHandler
     fun calculateStats(rqt: RequestFinalized) {
-        val (accNo, version) = rqt
-        logger.info { "$accNo, Received calculate status message for submission $accNo, version: $version" }
-        stagesHandler.calculateStats(rqt)
+        processSafely(rqt) {
+            val (accNo, version) = rqt
+            logger.info { "$accNo, Received calculate status message for submission $accNo, version: $version" }
+            statsService.calculateSubFilesSize(rqt.accNo)
+        }
+    }
+
+    private fun processSafely(request: RequestMessage, process: suspend RequestMessage.() -> Unit) = runBlocking {
+        runCatching { process(request) }.onFailure { onError(it, request) }
+    }
+
+    private fun onError(exception: Throwable, rqt: RequestMessage) {
+        logger.error(exception) { "${rqt.accNo}, Problem processing request '${rqt.accNo}': ${exception.message}" }
+        eventsPublisherService.submissionFailed(rqt)
     }
 }
